database_test-main/requirements.txt
alabaster==0.7.12
amqp==2.6.1
anyascii==0.2.0
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.4
astroid==2.5.6
async-generator==1.10
attrs==21.2.0
autopep8==1.5.5
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
biopython==1.78
black==21.5b1
bokeh==2.3.2
boto3==1.17.81
botocore==1.20.81
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
commonmark==0.9.1
coverage==5.5
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.13
django-allauth==0.44.0
django-celery-beat==2.2.0
django-ckeditor @ git+https://github.com/django-ckeditor/django-ckeditor.git@ae046476f7aea9d2b5f7d359ebe596cb6a175a37
django-crispy-forms==1.11.2
django-dbbackup==3.3.0
django-docs==0.3.1
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-modeladmin-reorder==0.3.1
django-modelcluster==5.1
django-ses==2.0.0
django-taggit==1.4.0
django-timezone-field==4.1.2
django-treebeard==4.5.1
djangorestframework==3.12.4
docutils==0.16
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
flake8==3.8.4
future==0.18.2
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
h11==0.12.0
html5lib==1.1
httplib2==0.19.1
idna==2.10
imagesize==1.2.0
importlib-metadata==4.2.0
ipaddress==1.0.23
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.17.2
Jinja2==3.0.1
jmespath==0.10.0
kombu==4.6.11
l18n==2020.6.1
lazy-object-proxy==1.6.0
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
natsort==7.1.1
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
outcome==1.1.0
packaging==20.9
pandas==1.2.4
parso==0.7.1
pathspec==0.8.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
pretty-errors==1.2.24
prompt-toolkit==3.0.18
protobuf==3.17.3
psycopg2==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.2.0
Pygments==2.9.0
PyJWT==2.1.0
pylama==7.7.1
pylint==2.8.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-http-client==3.3.3
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-size==0.1.1
python3-openid==3.2.0
pytz==2021.1
pyuca==1.2
pyudev==0.22.0
PyYAML==5.4.1
recommonmark==0.7.1
redis==3.5.3
regex==2021.4.4
reportlab==3.6.1
requests==2.25.1
requests-oauthlib==1.3.0
rich==10.2.2
rinoh-typeface-dejavuserif==0.1.3
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.5.3
rope==0.19.0
rsa==4.7.2
s3transfer==0.4.2
selenium==4.0.0
sendgrid==6.8.2
six==1.16.0
smartypants==2.0.1
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinx-celery==2.0.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
structlog==21.1.0
tablib==3.0.0
telepath==0.2
toml==0.10.2
tornado==6.1
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
wrapt==1.12.1
wsproto==1.0.0
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
yapf==0.31.0
zipp==3.4.1
